Add a "Seen_table" to track the meetings
This is needed to trach which teams have meeted and which not. The
implementation in python uses a dict of sets here, aopy a new version
of the dict on every recursion step and replaces sets that are changing.

This is a container that have a bitset for each team. The container is
copied for every recursion step and tracks the number of copies to
determine which bitset should be copied and which is already a copy for
the current generation.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>