Add old prototype of the algorithm.
This add the old algorithm that was used the last five or six events.
Its python-only and can run on pypy. On a average machine it can conly
handle less than 21 teams so we hat to cluster the field. But within
this range it works pretty good.

It was added for documentation purposes.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>