Add a wrapper for the Guest_tuple_generator
Boost::python requires operator* to return a reference to the current
element. Since the Guest_tuple_generator generates the elements lazy
and holds no reference to them this will not work and results in
unpredictable behaviout.

To have a python-mapping for testing the iterator this wrapper stores
all elements it generates in a list within the iterator and gibes
references to its elements. This is not perfect since the list and all
elements will be destroyed as soon as the iterator goes out of scope
but its good enough for testing.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>