Add some real calculation functions to Calculation
It has now two data-containers: one containg the round specific data and
one containing the iteration specific data. These containers are
structs because we want fast data access.

The random stuff is declared mutable because requesting random numbers
is not side-effect-free. But this should not matter because we need it
only to have some sort of random choicing at the beginning. There is no
need for a pefect distributed random. This allows us to declare a lot of
methods const

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>