Add a (temporary) wrapper for Calculation
This is needed because the Round_data_wrapper is a derived type of
Round_data but I was not able to tell this Boost.Python. It states that
the signatues does not match and quits service.

As soon as more functionality goes into the calculation we can drop
this eventually because there is no need to test single methods from
python.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>