Simplify Seen_table memory handling
As pointed out by @devreal a lot of time is lost by dynamic memory
management for the copy-on-write mechanism of the Seen_table. So this
commit simply drop this and use full copys of the data tables.

From am memory usage point of view this should not really matter,
because we remove the array of pointers to the table entries. A table
entry itself is obly a std::bitset of MAX_TEAMS - which takes up to
eight bytes. So we save eight bytes by not storing the pointers and use
another eight by having the whole table.

The table is mmecpy-ed - so we save the element copy-construction.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>